refactor: centralize shared automation components
