+++ src/components/automation/excutions/historical/data-table-row-actions.tsx
@@ -1,44 +1,1 @@
-'use client'
-
-import { Row } from '@tanstack/react-table'
-import { MoreHorizontal } from 'lucide-react'
-
-import { Button } from '@/components/ui/button'
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu'
-import type { ExcutionsRow } from '../excutions'
-
-interface DataTableRowActionsProps {
-  row: Row<ExcutionsRow>
-}
-
-export default function DataTableRowAction({ row }: DataTableRowActionsProps) {
-  const handleEdit = () => {
-    console.log('Edit asset:', row.original)
-  }
-
-  const handleDelete = () => {
-    console.log('Delete asset:', row.original)
-  }
-
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button variant="ghost" className="flex h-8 w-8 p-0 data-[state=open]:bg-muted">
-          <MoreHorizontal className="h-4 w-4" />
-          <span className="sr-only">Open menu</span>
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" className="w-[160px]">
-        <DropdownMenuItem onClick={handleEdit}>Edit</DropdownMenuItem>
-        <DropdownMenuSeparator />
-        <DropdownMenuItem onClick={handleDelete}>Delete</DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  )
-}
+export { SimpleRowActions as default } from '@/components/automation/common/simple-row-actions'
